Phase 4: decompose CourseModulesAdmin into cards, drafts module, and data hook
